fix:carousel 로직 분해
