+++ apps/web/.eslintignore
@@ -1,3 +1,4 @@
 tailwind.config.js
 postcss.config.js
 webpack.config.js
+eslint.config.js
